maps done

+++ app/src/main/java/com/example/storyapp/data/remote/responses/StoryResponse.kt
@@ -37,8 +37,8 @@ data class ListStoryItem(
 	val description: String,
 
 	@field:SerializedName("lon")
-	val lon: Float?,
+	val lon: Double?,
 
 	@field:SerializedName("lat")
-	val lat: Float?
+	val lat: Double?
 ) : Parcelable
